Move SimpleSparse from simple_block.py to jacobian.py
